requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Frontend
jinja2==3.1.2
# streamlit==1.28.1  # Not needed for production
# streamlit-option-menu==0.3.6  # Not needed for production

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# sqlite3 is built into Python

# AI/ML
openai==1.3.7
# sentence-transformers==2.3.1  # Removed - using only OpenAI APIs
numpy==1.24.3
# scikit-learn==1.3.0  # Removed - using only OpenAI APIs

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File handling
python-magic==0.4.27

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
